updater

+++ js/scripts.js
@@ -37,43 +37,49 @@ const increaseAmount = (event) => {
     Number(event.target.previousElementSibling.textContent) + 1;
   emptyCartElement.classList.add('empty-cart-hidden');
   itemCount.push(1);
-  console.dir(event.target);
-  const fragment = document.createDocumentFragment();
-  const itemOrderedAndPrice = document.createElement('div');
-  itemOrderedAndPrice.classList.add('item-ordered-and-prices');
-  const unitaryItem = document.createElement('div');
-  unitaryItem.classList.add('unitary-item');
-  const itemOrderedTitle = document.createElement('span');
-  itemOrderedTitle.classList.add('item-ordered-title');
-  itemOrderedTitle.textContent =
-    event.target.parentElement.parentElement.nextElementSibling.children[1].textContent;
-  const amountAndPrice = document.createElement('div');
-  amountAndPrice.classList.add('amount-and-price');
-  const amount = document.createElement('span');
-  amount.classList.add('amount');
-  amount.textContent = event.target.previousElementSibling.textContent + 'x';
-  const unitaryPrice = document.createElement('span');
-  unitaryPrice.classList.add('unitary-price');
-  unitaryPrice.textContent =
-    '@$' +
-    event.target.parentElement.parentElement.nextElementSibling.children[2]
-      .children[0].dataset.singleprice;
-  const itemCountPrice = document.createElement('span');
-  itemCountPrice.classList.add('price-per-product');
-  itemCountPrice.textContent =
-    '$' +
-    event.target.parentElement.parentElement.nextElementSibling.children[2]
-      .children[0].dataset.singleprice *
-      Number(event.target.previousElementSibling.textContent);
-  const deleteItem = document.createElement('div');
-  deleteItem.classList.add('delete-cart');
-  amountAndPrice.append(amount, unitaryPrice, itemCountPrice);
-  unitaryItem.append(itemOrderedTitle, amountAndPrice);
-  itemOrderedAndPrice.append(unitaryItem, deleteItem);
-  fragment.id =
-    event.target.parentElement.parentElement.nextElementSibling.children[1].dataset.dish;
-  fragment.append(itemOrderedAndPrice);
-  toAppendInCartElement.append(fragment);
+  if (itemCount.length === 1) {
+    const fragment = document.createDocumentFragment();
+    const itemOrderedAndPrice = document.createElement('div');
+    itemOrderedAndPrice.classList.add('item-ordered-and-prices');
+    itemOrderedAndPrice.dataset.dish =
+      event.target.parentElement.parentElement.nextElementSibling.children[1].dataset.dish;
+    const unitaryItem = document.createElement('div');
+    unitaryItem.classList.add('unitary-item');
+    const itemOrderedTitle = document.createElement('span');
+    itemOrderedTitle.classList.add('item-ordered-title');
+    itemOrderedTitle.textContent =
+      event.target.parentElement.parentElement.nextElementSibling.children[1].textContent;
+    const amountAndPrice = document.createElement('div');
+    amountAndPrice.classList.add('amount-and-price');
+    const amount = document.createElement('span');
+    amount.classList.add('amount');
+    amount.textContent = event.target.previousElementSibling.textContent + 'x';
+    const unitaryPrice = document.createElement('span');
+    unitaryPrice.classList.add('unitary-price');
+    unitaryPrice.textContent =
+      '@$' +
+      event.target.parentElement.parentElement.nextElementSibling.children[2]
+        .children[0].dataset.singleprice;
+    const itemCountPrice = document.createElement('span');
+    itemCountPrice.classList.add('price-per-product');
+    itemCountPrice.textContent =
+      '$' +
+      event.target.parentElement.parentElement.nextElementSibling.children[2]
+        .children[0].dataset.singleprice *
+        Number(event.target.previousElementSibling.textContent);
+    const deleteItem = document.createElement('div');
+    deleteItem.classList.add('delete-cart');
+    amountAndPrice.append(amount, unitaryPrice, itemCountPrice);
+    unitaryItem.append(itemOrderedTitle, amountAndPrice);
+    itemOrderedAndPrice.append(unitaryItem, deleteItem);
+    fragment.append(itemOrderedAndPrice);
+    toAppendInCartElement.append(fragment);
+  } else if (itemCount.length > 1) {
+    console.dir(toAppendInCartElement.children); //(
+    //      event.target.parentElement.parentElement.nextElementSibling.children[1]
+    //        .dataset.dish
+    //    ).children[0].children[1].children[0].textContent = itemCount.length;
+  }
 };
 allIncreaseElements.forEach((increaseElement) => {
   increaseElement.addEventListener('click', increaseAmount);
